# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Реактив CRM",
    "description": "CRM продаж и закупок химической продукции. Настраиваемые финансовые профили, права и шаблоны. Количества и деньги — десятичные строки.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chats": {
      "post": {
        "tags": [
          "Общение и файлы"
        ],
        "summary": "Create Chat",
        "operationId": "create_chat_api_v1_chats_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Chat Api V1 Chats Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      },
      "get": {
        "tags": [
          "Общение и файлы"
        ],
        "summary": "Chats",
        "operationId": "chats_api_v1_chats_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Chats Api V1 Chats Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/chats/{entity_id}": {
      "get": {
        "tags": [
          "Общение и файлы"
        ],
        "summary": "Get Chat",
        "operationId": "get_chat_api_v1_chats__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Chat Api V1 Chats  Entity Id  Get"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      },
      "patch": {
        "tags": [
          "Общение и файлы"
        ],
        "summary": "Edit Chat",
        "operationId": "edit_chat_api_v1_chats__entity_id__patch",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Edit Chat Api V1 Chats  Entity Id  Patch"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/chats/{entity_id}/members": {
      "post": {
        "tags": [
          "Общение и файлы"
        ],
        "summary": "Add Member",
        "operationId": "add_member_api_v1_chats__entity_id__members_post",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Add Member Api V1 Chats  Entity Id  Members Post"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
    "/api/v1/chats/{entity_id}/members/{member_id}": {
      "delete": {
        "tags": [
          "Общение и файлы"
        ],
        "summary": "Remove Member",
        "operationId": "remove_member_api_v1_chats__entity_id__members__member_id__delete",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Remove Member Api V1 Chats  Entity Id  Members  Member Id  Delete"
                }
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    },
...
    "/api/v1/files/{entity_id}/download": {
      "get": {
        "tags": [
          "Общение и файлы"
        ],
        "summary": "Download File",
        "operationId": "download_file_api_v1_files__entity_id__download_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Сеанс отсутствует или отозван"
          },
          "403": {
            "description": "Недостаточно прав"
          },
          "409": {
            "description": "Конфликт версии или повторяемости"
          },
          "422": {
            "description": "Ошибка предметной валидации"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChatInput": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 250,
            "minLength": 1,
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "enum": [
              "direct",
              "group",
              "request",
              "wave"
            ],
            "title": "Kind"
          },
          "member_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Member Ids"
          },
          "entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "kind"
        ],
        "title": "ChatInput"
      },
      "ChatPatch": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 250,
            "minLength": 1,
            "title": "Title"
          },
          "version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "version"
        ],
        "title": "ChatPatch"
      },
...
    }
  }
}